Add ArrowGuessBurrow and TimeLimitedList utilities
Introduces ArrowGuessBurrow for arrow particle detection and a generic TimeLimitedList utility for time-based element expiration. Moves PreciseGuessBurrow to the guesses package and updates WaypointManager import accordingly. Also adjusts guess timing logic in PreciseGuessBurrow.
