+++ src/main/kotlin/net/sbo/mod/utils/waypoint/WaypointManager.kt
@@ -2,7 +2,7 @@ package net.sbo.mod.utils.waypoint
 
 import net.fabricmc.fabric.api.client.rendering.v1.WorldRenderContext
 import net.fabricmc.fabric.api.client.rendering.v1.WorldRenderEvents
-import net.sbo.mod.diana.PreciseGuessBurrow
+import net.sbo.mod.diana.guesses.PreciseGuessBurrow
 import net.minecraft.client.MinecraftClient
 import net.minecraft.client.world.ClientWorld
 import net.minecraft.util.math.BlockPos
